+++ R/connectcreds-package.R
@@ -4,5 +4,6 @@
 ## usethis namespace: start
 #' @import httr2
 #' @import rlang
+#' @importFrom R6 R6Class
 ## usethis namespace: end
 NULL